New feature proposal - search by co-ordinates
