+++ lib/ip.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var _ = require('lodash');
-var Result = require('./result');
+var ResultBuilder = require('./result-builder');
 
 function IP(ip) {
   this.ip = ip;
@@ -9,20 +9,7 @@ function IP(ip) {
 }
 
 _.extend(IP.prototype, {
-  buildResult: function(body) {
-    body = body || {};
-
-    return new Result({
-      countryCode: body.country_code,
-      country: body.country_name,
-      regionCode: body.region_code,
-      region: body.region_name,
-      city: body.city,
-      postalCode: body.zip_code,
-      lat: body.latitude,
-      lng: body.longitude
-    });
-  }
+  buildResult: ResultBuilder._buildIpResult,
 });
 
 module.exports = IP;
